Fixed issue where times were being treated as ttc lines
This was caused by the link regex catching things and removing them,
which prevent other regexs from catching the things they needed to catch.

EX. "4 a.m.," was being caught by the link regex first and being set to
"4 ," which meant the time regex couldnt catch that it was a time. This
meant the "4" was left in to be treated as a ttc line.